+++ routes/users.js
@@ -1,24 +1,11 @@
 const express = require('express');
 const router = express.Router();
-const connection = require('../config/mysqlConfig');
-const { StatusCodes } = require('http-status-codes');
+const { signup } = require('../controllers/userController');
 
 router.use(express.json());
 
 // 회원가입
-router.post('/signup', (req, res) => {
-  const { email, password } = req.body;
-  const sql = `INSERT INTO users (email, password) VALUES (?, ?)`;
-  const values = [email, password];
-  console.log(email, password);
-
-  connection.query(sql, values, (err, results) => {
-    if (err) {
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.CREATED).json(results);
-  });
-});
+router.post('/signup', signup);
 
 // 로그인
 router.post('/signin', (req, res) => {
